fix ~ remove unneeded end-of-process color mode restoration code
.# Discussion

All output containing color sequences already shifts back to normal/default mode at the
end of the colored text output. So, the restoration code adds an unneeded bit of text
(and an additional line) to the output. This is especially noticeable and annoying when
the output is piped to another command (eg, `less`).